V1.13 Updates
A biggie here in terms of compatibility here back: I've redesigned the flywheel cage to allow it to be printed in the opposite of the original orientation: with the side that the motors insert into as the bottom layer. Less supports (though still recommended), and hopefully less alignment issues with warping and other print defects. This required making the cage 1.5mm thicker (and in order to minimize this extension, I reduced the bottom clearance for the flywheels by 1mm (don't worry, it was initially deliberately excessive), and the mounting points were moved up 2mm. Therefore, the magwell is also now slightly different, and so are all motor cover designs. A new page has been created for the older cage design and can be found in a remix. Another update is that I now feel comfortable releasing a foregrip design: it mounts onto the magwell, you'll need a new print for it (a magwell compatible with the old cages can be found in the legacy cage page too).