V1.2 Update
The next big-ish update as it includes a feature I've put off working on and testing since I completed the blaster: a full-length magwell. A new foregrip for that magwell comes with it, all picatinny options between one and two-stages are available. I kind of include the 'carbine' remix as part of this update, which is the main other feature.