V1.21 STL Files
This is a large update and one of the final ones. The new 'EXPD Battery Stock' is available in metric, imperial, kill switch and voltmeter variants, and has volume for up to 110x36x31mm packs (though it should be at least a little smaller for wiring), and assembles in the same way as the original stock. There is also a new 'Stock Cap (Attach)' option: instead of a printed sling mount, a Worker or very similarly mounted sling can be attached instead. Furthermore, there are now 14.5mm bore cages: occasional reports of inconsistent FPS results, the difficulty of pinpointing why has led me to make these an option instead: velocity (consistency) is just as important as precision to tagging players. The 38.5mm Eclipse option is not available as the wider bore begins to interfere with the cage geometry, and honestly I just didn't want to bother. Finally, I extended the end of the pusher by 1mm, it still doesn't protrude into the magwell but this should account for some reliability issues that have been reported with failing to feed into Daybreaks.