V1.21 Fixes
Files with the fixed flash hider mounts (to flywheel cover for 14.5mm cages) in CAD, including also hidden changes to the 'FPS dropper' attachment. It still is not in the STL files as I don't deem it fit, but it's there and some people seem to have gotten nicer groupings out of it.