V1.22 Update (03/07/21):
Developed MAX cages for single-stage variants. These are 15mm bore (essentially eliminating barrel drag), 39.5mm center-distance cages for Daybreak / Twilight wheels. As Eclipse wheels are not anywhere near available, this is a quick variant for maximizing FPS out of a single-stage. These have not been sorted into the ZIP yet until I get access to a Mac to fix the issues with ":" in the file names for x86 with the EXPD stocks. Bless for the pull request and again, sorry for taking so long to get to it.