V1.24 (05/01/21):
Added a side battery tray design to the files. This is something that's been developed over a few different remixes by different people, but I either didn't like the appearance of many of them, the implementation of the tray door, or the lack of space offered in most of those designs. The internal volume of the tray is about 110x35x24mm (you will of course, need to slightly undersize), it uses a captive M3 nut and thumbscrew + latch for tool-less opening, and has a sling mount on the side that replaces the stock cap.