V1.25 Update (05/30/21)
Altered the side battery tray design file: I wasn't happy with a lot of the first iteration. Firstly, I simplified the design to not require one to remove the top mech cover to install. Then, I added a mm of extra space in the dimension of a pack's width as I've found that the wire leads from a pack often add extra width or prevent fit. Finally, I created various tray door options ranging from a low-profile to an extended variant (for packs which take up all of the tray space and it can be impossible to fit an XT60 in the remaining space), and a low-profile variant with a cut-out should you want the low profile but not mind having dangling wires.