V1.3 Full-Length Magwell Changes
Part of the V1.3 update which primarily involves a fundamental restructuring of the pusher mech/magwell assembly. The mech covers have been merged into one piece and redesigned to print in such a way and retain space for wiring and guiding the pusher. The grip frame has been modified to be merged with a wider magwell joint, which should hopefully be more structurally sound in addition to reducing the part and hardware count. The Mech Cover to Grip attachment and Magwell to Grip Frame attachment points have been redesigned to accept M3x25 screws and an M3 hex nut to further drop hardware requirements. This cumulatively drops the part count by 3 and the screw count by 8.

All of the flywheel cages (not just the MAX cages now) have been re-exported with ribs in the motor wells to hold the motors more securely and lessen flexing from higher crushes, which should increase long-term durability and possibly marginally increase velocity for high envelopment/crush systems. Also enforced a uniform 1mm fillet in the wheel cavity-to-bore geometry (some, like the 40mm cages had a 0.5mm fillet previously).

Got rid of all the original picatinny options and enforced a more uniform standard: three splits of picatinny between the pusher mech, magwell, and flywheel cage. Allows for these parts of the blaster to be split up without undoing any picatinny and just reduces confusion between all the previous options of monolithic or etc. picatinny.

There is also a major additional feature/option: a buffer-tube mounted battery stock. The buffer tube is intended to be bonded with a 25mm diameter PVC inside to reinforce it (3/4 PVC is too thick for the walls to have any semblance thickness, so there is no imperial option here). For printability, the stock cap and buffer tube are printed separately (the former still requires support like any other stock cap), but the stock itself prints without support. The stock's battery holding internal dimensions measure at 107x36x26.8mm, and some of the wiring can be left in the buffer tube so a battery can be more closely spec'd to these dimensions. The KS and/or VM options were disappointing to me and with the adjustable length it can be extra finicky to set up a kill switch and/or voltmeter in the stock. I would personally recommend running without for wiring simplicity / and reducing rubbing/potential pinching.

Finally, tweaked the side battery tray to have a larger sling mount (it was far too small before) and adjusted the fit of the tray door's hook to be tighter.