V1.31 Update: Mech Cover
Removed hold-down geometry on the wire channels in the mech cover: wires can now just be slotted over instead of needing to be threaded through. I decided that with appropriate lubrication and thick enough wire cut to an appropriate length, there shouldn't be much issue with the pusher mech binding up.