V1.31 Update: Revised buffer tube design
The buffer tube was changed to print with the top of the tube as the bottom layer. In order to cover the top mount of the stock, an extra piece with the sling mount has been added (as the new print orientation would also weaken the sling mount if it couldn't be printed in a different orientation). This should reduce print complexity and eliminate the need to reinforce the buffer tube.